fix(i18n): echo bare-string worksheet entries; widen read guard
Re-review findings on the fix batch.

The stderr transcript promises "the discarded text is echoed here
— it is not saved anywhere else", and for one entry shape it
echoed `''`: a bare-string entry (`{"a.key": "Ein Text"}`) never
reached the dict branch. That is the likeliest hand-written
worksheet, and the very shape whose rejection made the
`unverifiable` bucket necessary — so the one case the mitigation
silently dropped was the one it was written for.

`_load_map` caught only `json.JSONDecodeError`, but
`read_text(encoding="utf-8")` raises `UnicodeDecodeError` first,
and the worksheet is the single file in this pipeline a
non-developer edits — an editor saving it as Latin-1 still dumped
a traceback. `OSError` joins them; the message already reads
correctly for all three.

A non-string `"translation"` was reported as an *empty* one,
which would send a translator back to look for work they had
already done. It moves to `unverifiable`, and that bucket's
message now names the not-in-worksheet-shape case explicitly
rather than leaving it under "no usable source".

`MergeKeysTests` crossed the 350-line ceiling at 381 in the
previous commit. Split at the existing seam: the guard's own
cases stay, everything else (malformed input, the pre-existing
empty/missing paths, `--site`) moves to `MergeKeysBehaviourTests`
— 256 and 253 lines. Its header records that its `--site` case is
the only `--site` coverage in the whole test tree, so a future
cleanup does not read it as redundant.

Four new tests, one per finding above.
