+++ Tests/KiwiDeskCoreTests/MergeKeysTests.swift
@@ -253,129 +253,4 @@ struct MergeKeysTests {
         #expect(worksheet?["gap.hint"]?["source"] == "New meaning")
         #expect(worksheet?["gap.hint"]?["translation"] == "")
     }
-
-    // MARK: - Pre-existing behaviour the guard must not break
-
-    @Test("an empty translation is still skipped, not merged")
-    func skipsEmptyTranslation() throws {
-        let fx = try fixture(locales: [
-            "en.json": #"{"gap.hint": "Meaning"}"#,
-            "de.json": #"{}"#,
-            "missing_de.json": #"""
-            {"gap.hint": {"source": "Meaning",
-                          "translation": "   "}}
-            """#,
-        ])
-        defer { fx.cleanup() }
-
-        let result = try run(["de"], in: fx)
-        #expect(result.status == 0)
-        #expect(try fx.decodeLocale("de.json").isEmpty)
-        #expect(result.stdout.contains("gap.hint"))
-    }
-
-    @Test("fails when the worksheet does not exist")
-    func failsWithoutWorksheet() throws {
-        let fx = try fixture(locales: [
-            "en.json": #"{"gap.hint": "Meaning"}"#
-        ])
-        defer { fx.cleanup() }
-
-        let result = try run(["de"], in: fx)
-        #expect(result.status != 0)
-        #expect(result.stderr.contains("extract-keys"))
-    }
-
-    @Test("fails when en.json is absent — nothing is verifiable")
-    func failsWithoutEnglish() throws {
-        let fx = try fixture(locales: [
-            "de.json": #"{}"#,
-            "missing_de.json": #"""
-            {"gap.hint": {"source": "Meaning",
-                          "translation": "Bedeutung"}}
-            """#,
-        ])
-        defer { fx.cleanup() }
-
-        let result = try run(["de"], in: fx)
-        #expect(result.status != 0)
-        // Refusing beats merging data no guard could check.
-        #expect(fx.localeFileExists("missing_de.json"))
-        #expect(try fx.decodeLocale("de.json").isEmpty)
-    }
-
-    @Test("creates the locale file when it does not exist yet")
-    func createsMissingLocaleFile() throws {
-        let fx = try fixture(locales: [
-            "en.json": #"{"gap.hint": "Meaning"}"#,
-            "missing_fr.json": #"""
-            {"gap.hint": {"source": "Meaning",
-                          "translation": "Sens"}}
-            """#,
-        ])
-        defer { fx.cleanup() }
-
-        let result = try run(["fr"], in: fx)
-        #expect(result.status == 0)
-        #expect(
-            try fx.decodeLocale("fr.json") == [
-                "gap.hint": "Sens"
-            ]
-        )
-    }
-
-    // MARK: - `--site`
-
-    @Test("--site reads and writes the site manifests")
-    func siteModeUsesSiteDirectory() throws {
-        let fx = try fixture(
-            // The app tree holds a *different* English for the
-            // same key, so passing --site against it would show
-            // up as a stale skip rather than a silent pass.
-            locales: [
-                "en.json": #"{"hero.title": "App English"}"#
-            ],
-            siteLocales: [
-                "en.json": #"{"hero.title": "Site English"}"#,
-                "de.json": #"{}"#,
-                "missing_de.json": #"""
-                {"hero.title": {"source": "Site English",
-                                "translation": "Seitentitel"}}
-                """#,
-            ]
-        )
-        defer { fx.cleanup() }
-
-        let result = try run(["--site", "de"], in: fx)
-        #expect(result.status == 0)
-        #expect(
-            try fx.decodeLocale("de.json", site: true) == [
-                "hero.title": "Seitentitel"
-            ]
-        )
-        // The app tree must be untouched.
-        #expect(!fx.localeFileExists("de.json"))
-    }
-
-    @Test("--site applies the stale-source guard too")
-    func siteModeGuardsStaleSource() throws {
-        let fx = try fixture(
-            locales: ["en.json": #"{"hero.title": "Whatever"}"#],
-            siteLocales: [
-                "en.json": #"{"hero.title": "New site copy"}"#,
-                "de.json": #"{}"#,
-                "missing_de.json": #"""
-                {"hero.title": {"source": "Old site copy",
-                                "translation": "Alt"}}
-                """#,
-            ]
-        )
-        defer { fx.cleanup() }
-
-        let result = try run(["--site", "de"], in: fx)
-        #expect(result.status == 0)
-        #expect(try fx.decodeLocale("de.json", site: true).isEmpty)
-        #expect(result.stderr.contains("hero.title"))
-        #expect(result.stderr.contains("--site"))
-    }
 }
